+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.55] - 25/03/2025
+
+### Added
+
+- Show warning when text input is larger that the current context window (Ollama)
+
 ## [0.1.54] - 25/03/2025
 
 ### Added